refactor(quran): extract complex karaoke word-sync and audio engine into reusable hook
